agent_writer: Renamed to agent_io
More fitting name, since it also deals with reading in the agents.
