+++ include/models/ActivityDrivenModel.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "agent.hpp"
-#include "agent_writer.hpp"
+#include "agent_io.hpp"
 #include "model.hpp"
 #include "network.hpp"
 #include <cstddef>
