+++ include/models/DeGroot.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "agent.hpp"
-#include "agent_writer.hpp"
+#include "agent_io.hpp"
 #include "model.hpp"
 #include "network.hpp"
 #include <vector>
